Error records now reset
Fixed a long-standing bug by which we did not ever erase ODBC error records.

All functions except for the functions that actually report errors are meant to erase
previous error records when they are invoked. I decided that the easiest/most clear
way to implement this was to simply copy/paste the reset line to each function.

An alternative would have been to add state, preferably in the handles themselves,
to keep track of the last record that was asked for by an error reporting function
and should errors.length ever exceed that value, to then remove the old entries.
This would likely be correct and does hide this implementation detail from the
ODBC interface functions, it also means adding a substantial wrapper around the
errors array. I am not opposed to doing this if it is preferable.

Couple of other things in this diff:
1: OdbcConnections now support errors
2: GetInfo had a TODO waiting for #1 to be done, that has been tended to