Compile on OSX
This is a *first* step. This does *not* complete the port to OSX.
The changes in this pull request simply make the driver compile, it
does not make it run properly. There are a few more details to work out:
1. The shared static this OSX equivalent of DllMain is currently *not*
executed. This means the runtime is never initialized and the driver
then crashes.
2. There are a few more usages of the long integer type that must be
removed/tended to. Ideally we would find out that there's a 64-bit
iODBC driver, but otherwise we need to deal with this.