Add SQLDriverConnect 'GUI'
When applications call SQLDriverConnect and ask us to prompt the user for more
information, we now pop open notepad as our GUI.

Renamed schema connection parameter to prestoSchema because Tableau assumes some
sort of undocumented purpose for 'schema' and fails to retain it in the connection
string. This fixed a long-standing point of confusion.

Added an alias to the 'database' connection parameter, 'prestoCatalog'.

Added another markdown file to describe the connection parameters and how to enter
them.

Future work: In the future I shall try to present a more user-friendly template
in notepad. For now, this should actually enable MS Query to work, I shall test that
and revise the readme in a near-future diff.